moving timer to util
